refactor(finance): 拆分房贷计算器数据模型，主算法文件瘦身解耦
- 新建 mortgage_models.py：常量、金额/日期工具、结果 dataclass 与序列化
- mortgage_calculator.py 通过 re-export 保持公共 API，外部导入不受影响
- 纯行为迁移，不改变任何计算结果（既有 6 个陈旧测试断言在拆分前已失败）
